refactor(cartridge): extract loader into _load.py
Move cartridge_to_preset, _resolve_extends, _shallow_merge_config, and
_workspace_to_preset_inner from looplet.cartridge.__init__ into a new
looplet.cartridge._load submodule. The old names remain importable
from looplet.cartridge via re-export.

__init__.py shrinks from ~1270 to ~210 lines and is now purely a
re-export façade: docstring + __all__ + imports of the symbols that
have moved into focused submodules (_layout, _yaml, _imports,
_resources, _render, _serialise, _manifest, _load). Behaviour is
unchanged.

Tests: 1941 pass, 1 skipped.
